requirements.txt
aiohappyeyeballs==2.3.2
aiohttp==3.10.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
coloredlogs==15.0.1
dataclasses-json==0.6.7
distro==1.9.0
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
faiss-cpu==1.8.0.post1
filelock==3.15.4
FlashRank==0.2.8
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.139.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
greenlet==3.0.3
groq==0.9.0
grpcio==1.65.2
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.24.5
humanfriendly==10.0
idna==3.7
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.11
langchain-community==0.2.10
langchain-core==0.2.25
langchain-google-genai==1.0.8
langchain-groq==0.1.8
langchain-text-splitters==0.2.2
langsmith==0.1.94
marshmallow==3.21.3
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
onnxruntime==1.18.1
orjson==3.10.6
packaging==24.1
proto-plus==1.24.0
protobuf==4.25.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.2
pypdf==4.3.1
pyreadline3==3.4.1
python-dotenv==1.0.1
PyYAML==6.0.1
rank-bm25==0.2.2
requests==2.32.3
rsa==4.9
sniffio==1.3.1
SQLAlchemy==2.0.31
sqlparse==0.5.1
sympy==1.13.1
tenacity==8.5.0
tokenizers==0.19.1
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
yarl==1.9.4